backend/requirements.txt
Flask
requests 
python-dotenv
flask_cors
google-generativeai